New Structure in repository
