+++ Prototypes/V2/Documentation/DocumentationV2.md
@@ -9,4 +9,3 @@ Motor Driver and perfboard in the back
 Custom 3D printed mounts for two Ultrasonic sensors and the Pixy 2.1 camera
 
 Cleaner cabling
-
